+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -418,6 +418,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a tooltip to the Ambient Mode parameter in the Visual Envionment volume component.
 - Static lighting sky now takes the default volume into account (this fixes discrepancies between baked and realtime lighting).
 - Fixed a leak in the sky system.
+- Removed MSAA Buffers allocation when lit shader mode is set to "deferred only".
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled